Renamed cryptographic_exception files to something more sensible
