+++ libs/cryptoplus/include/cryptoplus/asn1/integer.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_ASN1_INTEGER_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "../bio/bio_ptr.hpp"
 #include "../bn/bignum.hpp"
 
@@ -78,7 +78,7 @@ namespace cryptoplus
 				 * \brief Create a new integer.
 				 * \return The integer.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static integer create();
