+++ libs/cryptoplus/include/cryptoplus/asn1/object.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_ASN1_OBJECT_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 
 #include <openssl/objects.h>
 
@@ -74,7 +74,7 @@ namespace cryptoplus
 				 * \brief Create a new object.
 				 * \return The object.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static object create();
