+++ libs/cryptoplus/include/cryptoplus/asn1/string.hpp
@@ -47,7 +47,7 @@
 
 #include "../pointer_wrapper.hpp"
 #include "../buffer.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 
 #include <openssl/crypto.h>
 #include <openssl/asn1.h>
@@ -76,7 +76,7 @@ namespace cryptoplus
 				 * \brief Create a new string.
 				 * \return The string.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static string create();
