+++ libs/cryptoplus/include/cryptoplus/asn1/utctime.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_ASN1_UTCTIME_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "../bio/bio_ptr.hpp"
 
 #include <openssl/crypto.h>
@@ -77,7 +77,7 @@ namespace cryptoplus
 				 * \brief Create a new utctime.
 				 * \return The utctime.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static utctime create();
