+++ libs/cryptoplus/include/cryptoplus/bn/bignum.hpp
@@ -47,7 +47,7 @@
 
 #include "../pointer_wrapper.hpp"
 #include "../buffer.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 
 #include <openssl/bn.h>
 
@@ -79,7 +79,7 @@ namespace cryptoplus
 				 * \brief Create a new bignum.
 				 * \return The bignum.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static bignum create();
 
@@ -96,7 +96,7 @@ namespace cryptoplus
 				 * \param buf_len The length of buf.
 				 * \return A bignum.
 				 *
-				 * On error a cryptographic_exception is thrown.
+				 * On error an exception is thrown.
 				 */
 				static bignum from_bin(const void* buf, size_t buf_len);
