+++ libs/cryptoplus/include/cryptoplus/error/error.hpp
@@ -42,10 +42,13 @@
  * \brief Error helper functions.
  */
 
-#ifndef CRYPTOPLUS_ERROR_ERROR_HPP
-#define CRYPTOPLUS_ERROR_ERROR_HPP
+#pragma once
 
 #include <openssl/err.h>
+
+#include <boost/system/error_code.hpp>
+#include <boost/system/system_error.hpp>
+#include <boost/type_traits/integral_constant.hpp>
 
 namespace cryptoplus
 {
@@ -241,6 +244,61 @@ namespace cryptoplus
 			return ERR_GET_REASON(err);
 		}
 	}
+
+	/**
+	 * @brief Get the default cryptoplus error category.
+	 * @return The default cryptoplus error category instance.
+	 *
+	 * @warning The first call to this function is thread-safe only starting with C++11.
+	 */
+	const boost::system::error_category& cryptoplus_category();
+
+	/**
+	 * @brief Create an error_code instance for the given error.
+	 * @param error The error.
+	 * @return The error_code instance.
+	 */
+	inline boost::system::error_code make_error_code(error::error_type error)
+	{
+		return boost::system::error_code(static_cast<int>(error), cryptoplus_category());
+	}
+
+	/**
+	 * @brief Create an error_condition instance for the given error.
+	 * @param error The error.
+	 * @return The error_condition instance.
+	 */
+	inline boost::system::error_condition make_error_condition(error::error_type error)
+	{
+		return boost::system::error_condition(static_cast<int>(error), cryptoplus_category());
+	}
+
+	/**
+	 * @brief A cryptoplus error category.
+	 */
+	class cryptoplus_category_impl : public boost::system::error_category
+	{
+		public:
+			/**
+			 * @brief Get the name of the category.
+			 * @return The name of the category.
+			 */
+			virtual const char* name() const throw();
+
+			/**
+			 * @brief Get the error message for a given error.
+			 * @param ev The error numeric value.
+			 * @return The message associated to the error.
+			 */
+			virtual std::string message(int ev) const;
+	};
 }
 
-#endif /* CRYPTOPLUS_ERROR_ERROR_HPP */
+namespace boost
+{
+	namespace system
+	{
+		template <>
+		struct is_error_code_enum<cryptoplus::error::error_type> : public boost::true_type {};
+	}
+}
