+++ libs/cryptoplus/include/cryptoplus/hash/pbkdf2.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_HASH_PBKDF2_HPP
 
 #include "../buffer.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "message_digest.hpp"
 #include "message_digest_algorithm.hpp"
