+++ libs/cryptoplus/include/cryptoplus/x509/certificate.hpp
@@ -47,7 +47,7 @@
 
 #include "../pointer_wrapper.hpp"
 #include "../buffer.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "../bio/bio_ptr.hpp"
 #include "../pkey/pkey.hpp"
 #include "../asn1/integer.hpp"
@@ -219,7 +219,7 @@ namespace cryptoplus
 				 * \brief Create a new certificate.
 				 * \return The certificate.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static certificate create();
