+++ libs/cryptoplus/include/cryptoplus/x509/name_entry.hpp
@@ -47,7 +47,7 @@
 
 #include "../pointer_wrapper.hpp"
 #include "../buffer.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "../bio/bio_ptr.hpp"
 #include "../asn1/object.hpp"
 #include "../asn1/string.hpp"
@@ -78,7 +78,7 @@ namespace cryptoplus
 				 * \brief Create a new name_entry.
 				 * \return The name_entry.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static name_entry create();
