+++ libs/cryptoplus/include/cryptoplus/x509/store.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_X509_STORE_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "certificate.hpp"
 #include "certificate_revocation_list.hpp"
 #include "verify_param.hpp"
@@ -80,7 +80,7 @@ namespace cryptoplus
 				 * \brief Create a new store.
 				 * \return The store.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static store create();
