+++ libs/cryptoplus/include/cryptoplus/x509/store_context.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_X509_STORE_CONTEXT_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 
 #include "store.hpp"
 #include "certificate.hpp"
@@ -100,7 +100,7 @@ namespace cryptoplus
 				 * \brief Create a new store context.
 				 * \return The store context.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static store_context create();
