+++ libs/cryptoplus/include/cryptoplus/x509/verify_param.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_X509_VERIFY_PARAM_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 #include "../asn1/object.hpp"
 
 #include <boost/date_time/posix_time/posix_time.hpp>
@@ -73,7 +73,7 @@ namespace cryptoplus
 				 * \brief Create a new verify_param.
 				 * \return The verify_param.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static verify_param create();
