+++ libs/cryptoplus/include/cryptoplus/x509/x509v3_context.hpp
@@ -46,7 +46,7 @@
 #define CRYPTOPLUS_X509_X509V3_CONTEXT_HPP
 
 #include "../pointer_wrapper.hpp"
-#include "../error/cryptographic_exception.hpp"
+#include "../error/helpers.hpp"
 
 #include <openssl/x509v3.h>
 
@@ -71,7 +71,7 @@ namespace cryptoplus
 				 * \brief Create a new x509v3_context.
 				 * \return The x509v3_context.
 				 *
-				 * If allocation fails, a cryptographic_exception is thrown.
+				 * If allocation fails, an exception is thrown.
 				 */
 				static x509v3_context create();
